requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==3.6.4.0
celery==5.2.7
celery-singleton==0.3.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.2.0
cron-descriptor==1.4.3
cryptography==42.0.5
cycler==0.12.1
defusedxml==0.8.0rc2
Django==4.1.13
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-debug-toolbar==4.3.0
django-extensions==3.2.3
django-redis==5.2.0
django-timezone-field==6.1.0
flower==2.0.0
fonttools==4.49.0
humanize==4.9.0
idna==3.7
kiwisolver==1.4.5
kombu==5.3.5
matplotlib==3.8.3
numpy==1.26.4
oauthlib==3.2.2
packaging==24.0
Pillow==9.2.0
postgres==4.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
psycopg2-pool==1.2
pycparser==2.22
PyJWT==2.8.0
pyparsing==3.1.2
python-crontab==3.0.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2024.1
redis==5.0.3
requests==2.31.0
requests-oauthlib==2.0.0
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
tornado==6.4
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
djangorestframework==3.15.1
